Colocate API resources and serve canonical static OpenAPI docs
