+++ webapp/AGENTS.md
@@ -74,7 +74,7 @@
   - Tenant-owned tables must retain their own UUIDv7 ID, carry `organizationId` and `tenantId`, use all three columns as the primary key, and propagate composite organization/tenant foreign keys and unique constraints.
   - Better Auth's `member` and `invitation` tables are exceptions only to composite organization-scoped foreign keys; they still require UUIDv7 IDs and `organizationId` foreign keys. Obtain architectural approval for other exceptions.
   - Both audit columns use `DEFAULT now()`; Drizzle's `updatedAt` `$onUpdateFn` hook returns PostgreSQL `now()` but does not create a database trigger, so non-Drizzle updates must set `updated_at` explicitly.
-  - Define optimistic locks with `lockVersion()`, use `updateWithOptimisticLock` or `deleteWithOptimisticLock`, and map expected conflicts with `catchOptimisticLockConflict` instead of branching on error reasons at each caller; do not add `$onUpdateFn` or a redundant check constraint.
+  - Define optimistic locks with `lockVersion()`, use `updateWithOptimisticLock` or `deleteWithOptimisticLock`, and map expected conflicts with `catchOptimisticLockConflict` instead of branching on error reasons at each caller; do not add `$onUpdateFn` or a redundant check constraint. Tenant and TenantUser management APIs intentionally use last-write-wins without optimistic locks.
   - Store encrypted database values only through the strict Compact JWE helpers in `src/db/lib`. Required `DATABASE_ENCRYPTION_KEY` entries contain at least 32 characters: the first encrypts and the rest decrypt older values. Never treat malformed or unreadable JWE as plaintext, write during reads, log stored values, or expose them outside an explicitly authorized, no-store configuration flow. Include dynamic row identity inside the encrypted payload and compare it with sibling columns after decoding. Keep tenant-bound encrypted row access behind small table-specific functions. Use `src/db/config.server.ts` for validated reads and writes to the encrypted `config.value` column and `getGlobalConfig` for environment precedence and process caching; keep deployment settings global, organization settings authorized and organization-scoped, and use migrations instead of preemptive settings-schema version columns.
   - Surface sandbox provider configuration through the owner-and-developer `/:organizationSlug/sandboxes` sidebar entry and gate it by its exact Better Auth permission; keep every read and mutation protected independently at the server boundary because sidebar visibility is not authorization, mark authenticated configuration responses `Cache-Control: no-store`, and disable route preloading, cached loader retention, and stale rendering before returning decrypted credentials for masked editing.
   - Creating an organization also creates its starter agent (named after the organization) and the `organization_configuration` row whose `default_agent_id` points at it, so `/api/chat` can resolve an agent for a host that sends no `agentId`. Keep an agent's sandbox provider optional, because a provider cannot be saved before its connection test passes, and keep `default_agent_id` behind the composite organization-scoped reference, releasing it in the same transaction that deletes the agent it points at.
@@ -124,7 +124,7 @@
 - Public management REST APIs use the approved executable Effect HttpApi boundary; read `src/routes/api/v1/AGENTS.md` before changing those APIs, their repositories, OpenAPI export, or adding a subsequent API version. This does not replace dashboard server functions or chat transport.
 - Prefer supported framework or library APIs and well-maintained community packages over minor custom abstractions or bespoke integration plumbing.
 - Write new server-side application logic as Effect programs with typed failures; yield native Effect integrations directly, lift unavoidable Promise APIs with `Effect.tryPromise`, and run Effects only at framework boundaries.
-- Generate Effect validation schemas from Drizzle tables with the built-in [`drizzle-orm/effect-schema`](https://orm.drizzle.team/docs/effect-schema) helpers at database and API boundaries instead of duplicating table shapes by hand.
+- Generate Effect validation schemas from Drizzle tables with the built-in [`drizzle-orm/effect-schema`](https://orm.drizzle.team/docs/effect-schema) helpers at database and API boundaries instead of duplicating table shapes by hand. The versioned public REST API is an exception: keep its schemas independent of Drizzle so database changes cannot silently change its contract.
 
 - While writing DB queries, always put in the right authorization checks to avoid leaking one user/org's data to another
