+++ webapp/package.json
@@ -8,7 +8,7 @@
     "generate-routes": "tsr generate",
     "generate:openapi": "deno run -P=tooling scripts/export-openapi.ts",
     "openapi:check": "deno task generate:openapi --check",
-    "build": "NODE_ENV=production vite build",
+    "build": "deno task generate:openapi && NODE_ENV=production vite build",
     "start": "deno run --config deno.jsonc -P=runtime .output/server/index.mjs",
     "compile": "deno task typecheck && deno compile --no-check --node-modules-dir=none --exclude-unused-npm --include npm:busboy --config deno.jsonc -P=runtime --output .output/astralbeam .",
     "binary:check": "deno task compile && deno run --env-file=.env.test -P=tooling scripts/check-binary.ts",
@@ -47,6 +47,7 @@
     "@marsidev/react-turnstile": "^1.6.0",
     "@phosphor-icons/react": "^2.1.10",
     "@react-email/render": "2.1.0",
+    "@scalar/api-reference": "^1.67.0",
     "@tanstack/ai": "^0.49.1",
     "@tanstack/ai-openai": "^0.22.0",
     "@tanstack/ai-sandbox": "^0.5.1",
