+++ webapp/src/routes/api/v1/-lib/route.test.ts
@@ -1,4 +1,4 @@
-import { beforeEach, describe, expect, test, vi } from "vitest"
+import { afterEach, beforeEach, describe, expect, test, vi } from "vitest"
 import { getDatabaseBootstrapIssues } from "@/db/lib/database-credentials.server"
 import { setupGateResponse } from "@/lib/config/state.server"
 import { Route } from "../$"
@@ -11,6 +11,7 @@ const routeHandler = (Route.options.server!.handlers as {
 }).ANY
 
 describe("REST route setup boundary", () => {
+  afterEach(() => vi.restoreAllMocks())
   beforeEach(() => {
     vi.mocked(getDatabaseBootstrapIssues).mockReset().mockReturnValue([])
     vi.mocked(setupGateResponse).mockReset().mockResolvedValue(null)
@@ -46,5 +47,19 @@ describe("REST route setup boundary", () => {
     expect(response.headers.get("access-control-allow-credentials")).toBeNull()
     expect(getDatabaseBootstrapIssues).not.toHaveBeenCalled()
     expect(setupGateResponse).not.toHaveBeenCalled()
+  })
+  test("unexpected setup failures log safe diagnostics once", async () => {
+    const logged = vi.spyOn(console, "error").mockImplementation(() => {})
+    vi.mocked(setupGateResponse).mockRejectedValue(
+      Object.assign(new Error("private connection details"), { code: "08006" }),
+    )
+    const response = await routeHandler({ request: new Request("http://localhost/api/v1/tenants") })
+    expect(response.status).toBe(500)
+    expect(logged).toHaveBeenCalledExactlyOnceWith("API request failed", {
+      stage: "setup",
+      status: 500,
+      sqlState: "08006",
+    })
+    expect(await response.text()).not.toContain("private connection details")
   })
 })
