+++ webapp/src/routes/docs/-lib/scalar.server.ts
@@ -1,26 +1,25 @@
-import { HttpRouter } from "effect/unstable/http"
-import { HttpApiScalar, OpenApi } from "effect/unstable/httpapi"
-import { managementOpenApi, TenantRestApi } from "../../api/v1/-lib/contract.server"
-import gettingStarted from "../-content/api/getting-started.md?raw"
-import authentication from "../-content/api/authentication.md?raw"
-import pagination from "../-content/api/pagination-and-errors.md?raw"
+import type { ReferenceProps } from "@scalar/api-reference"
+// Scalar publishes its self-contained browser entry without a package subpath export.
+import scalarScript from "../../../../node_modules/@scalar/api-reference/dist/browser/standalone.js?raw"
 import appStyles from "../../../styles.css?raw"
 import scalarStyles from "./scalar.css?inline"
 import docsHeaderStyles from "./header.css?inline"
-import { APP_LOGO_LIGHT_SVG_URL } from "../../../lib/constants"
+import { APP_LOGO_LIGHT_SVG_URL, APP_NAME } from "../../../lib/constants"
 
 // Reuse the compiler's output without importing Tailwind resets or running the compiler at runtime.
 const apiDocsTheme = appStyles.split("/* Generated theme variables: start */")[1]!
   .split("/* Generated theme variables: end */")[0]!.replaceAll(".dark", ".dark-mode")
 
-export const apiDocsStyles = `${apiDocsTheme}\n${scalarStyles}\n${docsHeaderStyles}`
+const apiDocsStyles = `${apiDocsTheme}\n${scalarStyles}\n${docsHeaderStyles}`
 
-// Scalar accepts these options although Effect exposes only a subset of its configuration type.
 // https://scalar.com/products/api-references/configuration
 const apiDocsScalarOptions = {
   withDefaultFonts: false,
   hideClientButton: true,
   showDeveloperTools: "never",
+  documentDownloadType: "none",
+  agent: { disabled: true },
+  mcp: { disabled: true },
   authentication: {
     preferredSecurityScheme: "OrganizationApiKey",
     securitySchemes: {
@@ -56,37 +55,36 @@ const apiDocsScalarOptions = {
   },
   favicon: APP_LOGO_LIGHT_SVG_URL,
   customCss: apiDocsStyles,
-}
+} satisfies NonNullable<ReferenceProps["configuration"]>
 
-export const apiDocsHandler = HttpRouter.toWebHandler(
-  HttpApiScalar.layer(
-    TenantRestApi.annotate(
-      OpenApi.Description,
-      [gettingStarted, authentication, pagination].join("\n\n"),
-    ).annotate(OpenApi.Transform, apiDocsOpenApi),
-    { path: "/docs/api", scalar: apiDocsScalarOptions },
-  ),
-  { disableLogger: true },
-)
-
-// Group shared errors only in the human-facing reference; the public OpenAPI keeps explicit statuses.
-function apiDocsOpenApi(document: Record<string, unknown>): Record<string, unknown> {
-  const api = managementOpenApi(document) as unknown as OpenApi.OpenAPISpec
-  for (const methods of Object.values(api.paths)) {
-    for (const operation of Object.values(methods)) {
-      if (Array.isArray(operation)) continue
-      const responses: Record<string, OpenApi.OpenApiSpecResponse> = operation.responses
-      const error = responses["400"]
-      if (!error) continue
-      for (const status of Object.keys(responses)) {
-        if (Number(status) >= 400) delete responses[status]
-      }
-      responses.default = {
-        ...error,
-        description:
-          "Error response. See [Errors](#description/errors) for status codes and handling.",
+export function apiDocsHtml() {
+  return `<!doctype html><html><head>
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <title>${APP_NAME} API</title>
+    <style>${apiDocsStyles}</style>
+    </head><body>
+    <div id="api-reference-container"></div>
+    <script>${String(scalarScript).replaceAll("</script", "<\\/script")}</script>
+    <script type="module">
+    const response = await fetch('/api/openapi.json')
+    const content = await response.json()
+    // Group errors for presentation only; downloads keep the canonical explicit status codes.
+    for (const methods of Object.values(content.paths)) {
+      for (const operation of Object.values(methods)) {
+        const responses = operation.responses
+        if (!responses?.['400']) continue
+        responses.default = {
+          ...responses['400'],
+          description: 'Error response. See [Errors](#description/errors) for status codes and handling.',
+        }
+        for (const status of Object.keys(responses)) {
+          if (Number(status) >= 400) delete responses[status]
+        }
       }
     }
-  }
-  return document
+    Scalar.createApiReference('#api-reference-container', { ...${
+    JSON.stringify(apiDocsScalarOptions).replaceAll("<", "\\u003c")
+  }, content, baseServerURL: window.location.origin, generateTagSlug: ({ name }) => name })</script>
+    </body></html>`
 }
